Restart the explorer with the user account during un/installation
When restarting the explorer during installation and uninstallation,
previously the elevated installer just started the process as the
elevated user. This works if the currently logged-in user account is an
administrator and was used to elevate the installer, but it doesn't work
if the user is not an administrator and another account was used to
elevate the installer. In that case, the explorer simply wouldn't start
and the user was left without an explorer.

This fixes the problem by obtaining a reference to the user account from
the explorer process before closing it, and then later starting the
explorer again as that user.